requirements.txt
# FastAPI and Web Framework
fastapi>=0.115.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# LangGraph and LangChain
langgraph>=0.0.26
langchain>=0.1.0
langchain-google-genai>=0.0.6

# Google AI
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-cloud-speech>=2.0.0
google-cloud-texttospeech>=2.0.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
SQLAlchemy==2.0.25

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
